initial commit
created index.html and added structure of the doc. created two divs, one for display and one for the body of the calculater. inside the body, created four more divs for each line of the calculater. inside each line, added buttons as appropriate, for each button needed for the calculator.